% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collagen.R
\name{collagen}
\alias{collagen}
\title{Function to generate the call list with 1 click}
\usage{
collagen(raw)
}
\arguments{
\item{raw}{: the template dataframe obtained by the function collagen_read()}
}
\value{
'collagen_result-yyyy-mm-dd.csv' (the call list) and 'collagen_stat-yyyy-mm-dd.csv' (the statistics)
}
\description{
Function to generate the call list with 1 click
}
\examples{
raw = collagen_read()
collagen (raw)
}
