% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{collagen_read}
\alias{collagen_read}
\title{Function to read the collagen template}
\usage{
collagen_read()
}
\value{
A dataframe extracted from the collagen template
}
\description{
Function to read the collagen template
}
\examples{
collagen_read()
}
