% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{collagen_template}
\alias{collagen_template}
\title{Generating the collagen template for a specific month in a year}
\usage{
collagen_template(candidate_no, year, month, def_weekend)
}
\arguments{
\item{candidate_no}{: Number of candidates in the call list}

\item{year}{: Year (e.g. if the call list is for 2021-3 then year = 2021)}

\item{month}{: Month (e.g. if the call list is for 2021-3 then month = 3)}

\item{def_weekend}{: Weekend definitions, could be "SS" (Saturday and Sunday), "FS" (Friday and Saturday) and "FSS" (Friday, Saturday and Sunday)}
}
\value{
A collagen template in the working directory 'collagen_template-yyyy-mm-dd.csv'; yyyy-mm-dd refers to the date of generation, but not the concerned month for the call list
}
\description{
Generating the collagen template for a specific month in a year
}
\examples{
collagen_template(candidate_no = 12, year = 2021, month = 3, def_weekend = "SS")
}
