% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{gen_days}
\alias{gen_days}
\title{Finding the weekdays and weekends in a month}
\usage{
gen_days(year, month, def_weekend)
}
\arguments{
\item{year}{: Year (e.g. if the call list is for 2021-3 then year = 2021)}

\item{month}{: Month (e.g. if the call list is for 2021-3 then month = 3)}

\item{def_weekend}{: Weekend definitions, could be "SS" (Saturday and Sunday), "FS" (Friday and Saturday) and "FSS" (Friday, Saturday and Sunday)}
}
\value{
a list of characters signifying weekdays "D" and weekends "E" in a month
}
\description{
Finding the weekdays and weekends in a month
}
\examples{
gen_days(year = 2021, month = 3, def_weekend = "SS")
}
