% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collagen.R
\name{gen_PH}
\alias{gen_PH}
\title{Function to allocate the "P" dates ("P"ublic holiday or "P"ersonalized dates)}
\usage{
gen_PH(result, count, wecount, day, raww, raw)
}
\arguments{
\item{result}{: a variable holding the temporary results}

\item{count}{: a variable holding the temporary counts (total number of calls to be allocated for each candidates)}

\item{wecount}{: a variable holding the temporary weekend counts (total number of weekend calls to be allocated for each candidates)}

\item{day}{: a variable signifying weekdays "D" and weekends "E" and PH "P" in a month}

\item{raww}{: a simplified dataframe extracted from the collagen template}

\item{raw}{: the template dataframe obtained by the function collagen_read()}
}
\value{
a list of variables (result, count, wecount, day) for subsequent iterations
}
\description{
Function to allocate the "P" dates ("P"ublic holiday or "P"ersonalized dates)
}
\examples{
c(result, count, wecount, day) \%<-\% gen_PH(result, count, wecount, day, raww, raw)
}
