% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collagen.R
\name{gen_WD}
\alias{gen_WD}
\title{Function to allocate weekday calls}
\usage{
gen_WD(result, count, wecount, day, raww, raw)
}
\arguments{
\item{result}{: a variable holding the temporary results}

\item{count}{: a variable holding the temporary counts (total number of calls to be allocated for each candidates)}

\item{wecount}{: a variable holding the temporary weekend counts (total number of weekend calls to be allocated for each candidates)}

\item{day}{: a variable signifying weekdays "D" and weekends "E" and PH "P" in a month}

\item{raww}{: a simplified dataframe extracted from the collagen template}

\item{raw}{: the template dataframe obtained by the function collagen_read()}
}
\value{
a list of variables (result, count, wecount) for subsequent iterations
}
\description{
Function to allocate weekday calls
}
\examples{
c(result, count, wecount) \%<-\% gen_WD(result, count, wecount, day, raww, raw)
}
