% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collagen.R
\name{set_up_var}
\alias{set_up_var}
\title{Function to set up the variables from the template dataframe}
\usage{
set_up_var(raw)
}
\arguments{
\item{raw}{: the template dataframe obtained by the function collagen_read()}
}
\value{
a list of variables (result, count, wecount, day, raww) for subsequent iterations
}
\description{
Function to set up the variables from the template dataframe
}
\examples{
c(result, count, wecount, day, raww) \%<-\% set_up_var(raw)
}
